refactor: List<Integer> WinningNumbers 일급 컬렉션 사용
